pyproject.toml
[tool.poetry]
name = "base-python-boilerplate-backend"
version = "0.1.0"
description = ""
authors = ["Hugo Seabra <hugoseabra19@gmail.com>"]
readme = "README.md"
packages = [{include = "base_python_boilerplate_backend"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.5"
psycopg2-binary = "^2.9.5"
python-decouple = "^3.6"
psutil = "^5.9.4"
django-health-check = "^3.17.0"
django-redis = "^5.2.0"
celery = {extras = ["redis"], version = "^5.2.7"}
django-celery-results = "^2.4.0"
django-extensions = "^3.2.1"
docutils = "^0.19"
uuid = "^1.30"
djangorestframework = "^3.14.0"
markdown = "^3.4.1"
django-filter = "^22.1"
inquirer = "^3.1.1"
drf-nested-routers = "^0.93.4"
django-rest-swagger = "^2.2.0"

[tool.poetry.group.dev.dependencies]
werkzeug = "^2.2.2"
pyparsing = "^3.0.9"
pygraphviz = "^1.10"
pydot = "^1.4.2"
django-debug-toolbar = "^3.8.1"
faker = "^15.3.4"
coverage = "^7.0.1"
pyyaml = "^6.0"
uritemplate = "^4.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
